requirements.txt
# Core FastAPI
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0

# ML & Embeddings
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0
torch>=2.0.0
transformers>=4.30.0

# NLP & Text Processing
nltk>=3.8
spacy>=3.5.0
rank-bm25>=0.2.2

# Document Processing
pypdf>=3.17.0

# Visualization & Plotting
plotly>=5.14.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Dimensionality Reduction
umap-learn>=0.5.3
scikit-learn>=1.3.0

# PDF Export
jinja2>=3.1.2
weasyprint>=59.0

# Database
pydantic>=2.0.0

# Utilities
sh
